--- a/lib/stack/src/map_target.rs
+++ b/lib/stack/src/map_target.rs
@@ -1,3 +1,6 @@
+use futures::Poll;
+use svc;
+
 pub fn layer<T, M>(map_target: M) -> Layer<M>
 where
     M: MapTarget<T>,
@@ -8,7 +11,7 @@ where
 pub trait MapTarget<T> {
     type Target;
 
-    fn map_target(&self, t: &T) -> Self::Target;
+    fn map_target(&self, t: T) -> Self::Target;
 }
 
 #[derive(Clone, Debug)]
@@ -20,16 +23,10 @@ pub struct Stack<S, M> {
     map_target: M,
 }
 
-impl<T, S, M> super::Layer<T, M::Target, S> for Layer<M>
-where
-    S: super::Stack<M::Target>,
-    M: MapTarget<T> + Clone,
-{
-    type Value = <Stack<S, M> as super::Stack<T>>::Value;
-    type Error = <Stack<S, M> as super::Stack<T>>::Error;
-    type Stack = Stack<S, M>;
+impl<S, M: Clone> tower_layer::Layer<S> for Layer<M> {
+    type Service = Stack<S, M>;
 
-    fn bind(&self, inner: S) -> Self::Stack {
+    fn layer(&self, inner: S) -> Self::Service {
         Stack {
             inner,
             map_target: self.0.clone(),
@@ -37,25 +34,30 @@ where
     }
 }
 
-impl<T, S, M> super::Stack<T> for Stack<S, M>
+impl<T, S, M> svc::Service<T> for Stack<S, M>
 where
-    S: super::Stack<M::Target>,
+    S: svc::Service<M::Target>,
     M: MapTarget<T>,
 {
-    type Value = S::Value;
+    type Response = S::Response;
     type Error = S::Error;
+    type Future = S::Future;
 
-    fn make(&self, target: &T) -> Result<Self::Value, Self::Error> {
-        self.inner.make(&self.map_target.map_target(target))
+    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
+        self.inner.poll_ready()
+    }
+
+    fn call(&mut self, target: T) -> Self::Future {
+        self.inner.call(self.map_target.map_target(target))
     }
 }
 
 impl<F, T, U> MapTarget<T> for F
 where
-    F: Fn(&T) -> U,
+    F: Fn(T) -> U,
 {
     type Target = U;
-    fn map_target(&self, t: &T) -> U {
+    fn map_target(&self, t: T) -> U {
         (self)(t)
     }
 }