--- a/src/lib.rs
+++ b/src/lib.rs
@@ -29,9 +29,10 @@ extern crate rand;
 extern crate regex;
 extern crate tokio;
 extern crate tokio_timer;
+extern crate tower;
 extern crate tower_grpc;
 extern crate tower_http_service;
-extern crate tower_retry;
+extern crate tower_util;
 extern crate try_lock;
 
 #[macro_use]