--- a/src/proxy/http/mod.rs
+++ b/src/proxy/http/mod.rs
@@ -1,5 +1,6 @@
 pub mod add_header;
 pub mod balance;
+pub mod canonicalize;
 pub mod client;
 pub(super) mod glue;
 pub mod h1;