This converts from the previously used `linkerd2-stack` to `tower-layer`. This means that *most* things that were `Stack<T>` have been changed to `Service<T>`.

Some specific things changed to support the new pattern:

- The `linkerd2-router` would hold a `Stack<T>` to build services when a new target was recognized. The change in this PR is to create a specific `trait Make<T>` in `linkerd2-router` which can make a service immediately. In order to support this with the various `router::layer`s, a `pending::layer` has been created, which creates an `impl Service<Req>` immediately from a `MakeService<T, Req>`. The `Pending` service returns `NotReady` while the `MakeService::Future` is still pending. Once it completes, it acts just like the underlying service.
- `proxy::canonicalize` was holding onto a `Stack<T>` internally, and would build a new stack whenever DNS resolved to a new address. In to not need to add a buffer + pending layer just to support this, it was made a little more specific. It's moved to `proxy::http::canonicalize`, because it will insert the refined `Addr` into the `http::Request` extensions directly. Previously, since it held a `Stack<T>`, we had a tiny stack that was essentially just `insert_target`.