@olix0r 2019-04-23
now:
```
    Finished release [optimized] target(s) in 6m 23s
CARGO_RELEASE=1 make clean build  1082.30s user 33.68s system 288% cpu 6:26.26 total
```

master:
```
    Finished release [optimized] target(s) in 5m 28s
CARGO_RELEASE=1 make clean build  1013.46s user 29.43s system 316% cpu 5:29.04 total
```

It's interesting that this seems to have a substantial impact on build time. I wonder why?