@seanmonstar 2019-04-23
Oh that's interesting, I timed now and master but not in `--release` mode, and it was only a couple seconds difference, which I just assumed was noise.