@olix0r 2019-04-23
Seems about 15% hit in release mode, and about 25% in dev mode, if my math is right:

tower-layer:
```
    Finished dev [unoptimized] target(s) in 3m 40s
make clean build  396.19s user 25.32s system 190% cpu 3:41.68 total
```

master:
```
    Finished dev [unoptimized] target(s) in 2m 45s
make clean build  340.30s user 29.89s system 213% cpu 2:53.30 total
```