@olix0r 2019-04-23
That test passes without issue on master:
```
running 1 test
test inbound_http1 ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 2 filtered out
```