@olix0r 2019-04-23
Okay, `make test` succeeds; so I'm happy to dismiss the test failures to flakiness.